chore(scripts): 清理一次性脚本并扩展代码生成流水线
- 删除 frontend/scripts 与 scripts 下迁移/审计/批量修复等临时脚本

- 拆分 Vue 代码生成（crud/tree/master-detail/all）并保留 generate-all 主流水线

- 同步统计报表可配置、文件上传及全栈 DTO/规则更新

+++ backend/src/Takt.Application/Services/Identity/TaktAuthService.cs
@@ -875,7 +875,7 @@ public class TaktAuthService : TaktServiceBase, ITaktAuthService
         var roles = await _roleRepository.GetListAsync(r =>
             roleIds.Contains(r.Id) && r.TenantCode == tenantCode);
         return roles
-            .Where(r => r.RoleStatus == (int)TaktCommonStatus.Enabled)
+            .Where(r => r.RoleStatus == TaktCommonStatus.Enabled)
             .ToList();
     }
 
@@ -888,7 +888,7 @@ public class TaktAuthService : TaktServiceBase, ITaktAuthService
     private async Task<List<TaktMenu>> GetAccessibleMenusAsync(long userId, string tenantCode)
     {
         var allMenus = await _menuRepository.GetListAsync(m =>
-            m.TenantCode == tenantCode && m.MenuStatus == (int)TaktCommonStatus.Enabled);
+            m.TenantCode == tenantCode && m.MenuStatus == TaktCommonStatus.Enabled);
 
         var roles = await GetEnabledRolesAsync(userId, tenantCode);
         if (roles.Count == 0)
